source2il: implement a simple allocator (#92)
## Summary

Implement a heap allocator for internal use by `source2il`, for
implementing language features that require dynamic allocations (e.g.,
dynamic arrays).

## Details

The allocator is a simple bump-pointer allocator backed by a static
memory region. It consists of three procedures: `alloc`, `dealloc`, and
`realloc`. Since the source language doesn't support the features
necessary for implementing the allocator, the aforementioned procedures
are implemented in `L30` code provided via inline strings.

`dealloc` is currently a no-op, meaning that memory is never freed.
Upon exhausting the available heap memory, an exception is raised,
which currently results in undefined behaviour (as `CheckedCall` isn't
used for call expressions).

To reduce implementation complexity, the allocator procedures are
always added to the IL module, even if not used. Since this interferes
with the previous entry procedure selection (the last procedure in the
IL module was picked), all user-defined procedures are now marked as
exported, and for programs originating from source-language code, the
last *exported* procedure is now picked to be the entry procedure.

Finally, to be able to configure the guest memory from the outside via
modules, instead of it being hardcoded, the `phy` program now retrieves
the stack size, stack start, and total amount of memory from globals
with the interface names `stack_size`, `stack_start`, and
`total_memory`, respectively. This also means that the VM module is no
longer linked when not evaluating the program.