pass0: fix `BitNot` for signed integers (#101)
## Summary

Fix the translation of `BitNot` to bytecode resulting in incorrect
results for non-64-bit signed integers.

## Details

The result of the `opcBitNot` instruction was always masked, even for
signed integers, which is incorrect, as it invalidates the two's
complement representation. For example, `not 2'u32` should be -3,
but with the subsequent 32-bit masking it was 4,294,967,293‬
(`0xFF_FF_FF_FD`).

A mask instruction is now only emitted for unsigned integers. In
addition, an unnecessary but otherwise harmless `opcMask` instruction
was emitted for both `BitXor` (uint) and `Shl` (int), which is also
fixed.