source2il: insert copy calls for non-trivially-copyable values
Copying a non-trivially-copyable value requires a call to the
synthesized copy procedure. Except for the single lvalue-context usage
part of the `Asgn` translation, all `inline` usages are replaced with
the using the new `use`, which does the same as what `inline` does plus
emitting a call to the type's copy procedure (if needed for the source
expression).

For assignments, the destination's value also needs to be destroyed.