source2il: defer IL procedure list construction
Due to type-attached operator synthesis, it's now possible for
procedures to be constructed while construction of another procedure is
still in progress, in which case appending directly to the final IL
procedure list results in ill-formed modules.

To get around this, the IL procedure definitions are first accumulated
in a seq (so that starting construction of a procedure can reserve a
slot), with the seq elements then appended to the module when closing
the module.