lang: add built-in dynamic arrays (#79)
## Summary

Add run-time sized arrays with value semantics to the source language.
They're of type `SeqTy`, are constructed with `Seq` expressions, are
concatenated with `concat`, and have their dynamic length queried with
`len`.

## Details

Dynamic arrays are implemented as a length + pointer tuple, with the
pointer pointing to a dynamically allocated payload storing a capacity
(an integer) and an array of elements. Each use of a dynamic array
creates a full copy, which means creating a duplicate of the payload
and all elements within.

Growing a dynamic array and appending to it are implemented via
always-present type-erased L30 procedures. The copy procedure for types
needing one (queried via the new `isTriviallyCopyable` procedure) are
synthesized and cached when first needed. Calls to the copy procedure
are injected by the `use` procedure, which replaces all `inline`
usage in contexts where the specification describes a `copy` taking
place.

To support multiple procedures being constructed concurrently (needed
for ad-hoc type-attached operator synthesis), IL procedure definitions
are now first added to an intermediate seq before being added to the IL
module upon closing the analysis context.

Dynamic array values could and should be destructed when no longer
needed, but since destruction cannot be observed via in-language means
at the moment, it cannot be tested properly and is thus not
implemented.

Finally, `source2il` now keeps track of the most recently added user-
declared procedure's index in the `entry` field. With the synthesized
operators, the last procedure in the module's list is no longer
guaranteed to be a user-declared one, so `phy` and `repl` module use
the `entry` field to know which procedure to use as the entry
procedure.