skully: finish partial procedure support
Appending the fragments directly to the body of the partial
procedure - like the upstream backends do - is quite tricky to do with
skully. Therefore, each fragment is wrapped in its own procedure
instead. While a lot simpler, it does invalidate the 1-to-1 mapping
between MIR and IL procedure IDs, and thus a lookup table is required
now.