lang: add integer multiplication, division, and modulo (#103)
## Summary

Add the `*`, `div`, and `mod` built-in procedures for integers to the
source language. Integer addition and subtraction also work according
to the specification now (overflow leads to a panic/abort).

## Details

* add the static and dynamic semantics for integer multiplication,
  division, and modulo to the language specification
* translate `+` and `-` to `AddChck`/`SubChck` + `Unreachable` in
  `source2il` 
* add the `opcMulChck` opcode to the VM, for overflow-checked
  multiplication. The implementation is taken from Nim's compiler
  runtime library
* add the implementation `*`, `div`, and `mod` to `source2il` (with
  overflow / zero divisor checks)
* add tests for the add/sub overflow behaviour, as well as tests for
  the new operators 

Same as addition and subtraction, multiplication and division also
panic in case of overflow. Modulo uses truncated division because both
NimSkull and C (since C99) implement their modulo operator that way.
The `/` symbol is not used for division, so that it stays reserved for
a different operation (e.g., non-truncating division) in the future.