lang: add string constructors (#104)
## Summary

Add string constructors (`(Seq "...")`) and the `CharTy` type to the
source language. Both are meant as an interim solution to making
working with text easier.

## Details

* add the static and dynamic semantics for string constructors. A
  "character" is considered a UTF-8 byte, with the `(Seq "...")`
  constructing a dynamic array with the UTF-8 bytes of the given
  unicode string
* S-expression strings are automatically translated to
  `(StringVal ...)` during parsing
* add the `tkChar` type kind and implement the associated logic (type/
  value rendering, IL translation)

A `(StringVal ...)` is a static array (i.e., an array of statically
known size), and thus it's not considered to be a string constructor
itself.