lang: add `write` and `writeErr` (#105)
## Summary

Add support for writing to the standard output and error output to the
source language, via the built-in `write` and `writeErr` procedures.

## Details

For convenience, both output procedures take a `(SeqTy (CharTy))` value
as input. Internally, the built-ins invoke their corresponding external
procedure, which takes a data pointer and length as input.

Same as the allocator procedures, both new built-in procedures are
always added to a module, even if not used.

The operational semantics of both output procedures is modeled as
appending to an array stored in the context.