lang: add array indexing (#89)
## Summary

Add the `At` operator to the source language, for accessing elements of
arrays. All array access is checked, with out of bounds access
resulting in a panic.

## Details

* add the static and dynamic semantics for `At` to the specification
  * the interal-only `With` abstract syntax is introduced, to make
    defining the reduction of field/element assignments easier 
  * some small problems (incorrect notation, unintended meaning) with
    unrelated rules are fixed
  * the field access semantics is modified such that it matches the
    old one (in theory, no specification tests exist yet)

* implement the `At` syntax in `source2il`. Computing the element's
  address and checking the index needs to be an atomic expression, and
  thus the combined access/check is done via a call to procedure, which
  is synthesized for a `Seq` type upon first `At` use

* add tests for `At` and improve the existing `t17_seq_copy.test`

### Evaluation Order

The index operand may be an arbitrary effectful expressions, and
evaluating it can thus replace the dynamic array value in the location
that appears in the array operand position.

To prevent use-after-free errors without copying, the new evaluation
order for lvalue expressions is not strictly left to right. Instead,
all non-lvalue operands in the lvalue expression are evaluated first,
and only afterwards is the lvalue expression evaluated.