lang: add the `readFile` built-in (#107)
## Summary

Add the built-in `readFile` procedure to the source language, for
reading characters/bytes from arbitrary files.

## Details

* add the semantics for `readFile`. The content of external files is
  unknown, and thus `readFile` has an indeterminate result
* add the internal core API `readFile` and `fileSize` procedure and
  their implementations as VM host procedures
* implement the source language `readFile` as a wrapper around the core
  API `fileSize` and `readFile` procedures

In order to keep language semantics and implementation easier, at least
for the moment, attempting to read from a file that doesn't exist or
that is inaccessible for other reasons results in `readFile` returning
an empty sequence, rather than aborting the program.