add the `skully` program (#70)
## Summary

Skully (a portmanteau of "Skull" and "Phy") is a backend for NimSkull
that targets the `L25` language. It's combined with the NimSkull
frontend into a standalone program that takes a project module as input
and writes the generated `L25` code to a specified file.

Most NimSkull core language features are supported. The idea with
skully is to get access to large amounts of `L25` code resulting from
real-world code, for the purpose of testing and measuring the lowering
passes and VM.

## Details

The implementation is split into two parts:
* the actual code generator (`backend`) implementing the MIR to `L25`
  translation
* the program (`skully`) combining the NimSkull front- and mid-end with
  the Skully code generator

The `L25` IL is chosen as the target language because it's the language
most similar to the MIR.

Except for C interop, the `L25` is able to support the same language
features as NimSkull's C backend, and therefore `sem` is configured as
if it were run prior to the C backend. A non-trivial configuration is
used to make compilation possible:
* the target platform is hardcoded to be `any`, disabling most
  platform-specific code
* various internal conditional symbols are enabled to disable as much
  C-reliant code as possible
* a hook mechanism is used to replace importc'ed procedures with
  pure-NimSkull implementations or redirect them to dedicated VM host
  procedures
* platform-specific modules are replaced with skully-provided modules
  (only `os` at the moment, but the facility is generic and can
  redirect any module)

The VM host procedure specific to Skully are implemented into `phy`, so
that the `phy` program can run `skully`-produced code.

### Code Generator

The code generator performs a translation from MIR code to `L25` code,
attempting to replicate all lowerings NimSkull's C code generator
performs, meaning that the behaviour of the generated code *should* be
identical.

It's hooked into the compilation process like the native backends are:
via the `backends.process` iterator.

RTTIv1 generation is complicated and only needed for `deepCopy` and
the `typeinfo`, therefore it's not implemented by Skully; only the
RTTIv2 is, as it's necessary for ORC and `of` support.

### Platform Support

Only the OS procedures necessary for compiling the `phy` program are
implemented at the moment.

### Tests

For making sure Skully works at least somewhat, a CI job is added that
compiles `phy` to `L25` code and then executes said `L25` code to build
and run an more complex test from the `expr` tests.