implement a revised meta-language
The new meta-language is more powerful and also more generally
applicable, not enforcing a clear separation (or any separation at all)
between static and dynamic semantics. Not everything is implemented and
the meta-language is unlikely to be final, but it's a good start.

The source language definition using the old meta-language is removed.