source2il: fix missing seq copies (#111)
## Summary

Fix `SeqTy` values not being copied upon being passed to union and
tuple constructors, resulting in modifications of the constructed value
reflecting on the original sequence (and vice versa).

## Details

The `TupleCons` (tuple constructor) translation and union construction
(in `fitExpr`) emitted a direct use of their operands, ignoring whether
the type is trivially copyable or not.

`use` - which emits a call to the type's copy procedure where needed -
is now used in both cases, fixing the bug. A raw usage of `Expr.expr`
part of `While` translation is - while not causing any issues - also
replaced with a call to `use`.

The tests for sequence copies are expanded to also cover tuple,
union, and seq constructors.