spec: replace the type constructor shorthands
Summary
=======

Replace the "type shorthands" (`"int"`, `"float"`, etc.) with the
real syntax (`IntTy()`, `FloatTy(),` etc.) in the formal definition.

Details
=======

The type shorthands were originally intended to make the definition
easier to read, but they're currently defined as if they were separate
syntatic constructs (which they're not).

They're replaced with the type syntax they stand for, bringing the
language defintion closer to what was/is intended.