spec: replace the type constructor shorthands (#116)
## Summary

Replace the type shorthands (`"int"`, `"float"`, etc.) with the
real syntax (`IntTy()`, `FloatTy(),` etc.) in the formal definition.

## Details

The type shorthands were originally intended to make the definition
easier to read, but they're currently defined as if they were separate
syntactic constructs (which they're not).

They're replaced with the type syntax they stand for, bringing the
language definition closer to what was/is intended.