rework the meta language (#114)
## Summary

* turn the meta language into a more proper programming language 
* implement a type checker for the language (integrated into the macro)
* update the source language definition to use the new meta language

## Details

The reworked meta language is the result of attaching a type system to
it while keeping as many features and overall ergonomics as possible,
for the purpose of adding a static type checker.

The main changes are that:
* the language is less focused on syntax and pattern-matching
* it has a type system now and is statically typed
* it has more well-defined semantics, especially around non-pattern
  expressions

In more concrete terms:
* `nterminal` is replaced with `typ`, `subtype`, and `context`, with
  the former two representing inductively-defined abstract data types
  rather than s-expression-based syntax patterns
* formatting strings (on relation definitions) are removed
* support for negation and back-references in patterns is removed
* relations with only input parameters are usable like normal functions
* the confusing `condition`/`where`/`premise` usage in functions is
  replaced with `case` and `if` expressions and `let` statements

As a result, the meta language is now less like a description language
and more like a programming language. Instead of a *formal definition*,
a meta-language module now represents the *reference implementation* of
a language. Due to the simple structure and core semantics, an
independent formal definition can then be extracted from such
implementation, though this extraction is not yet implemented.

In addition to the `language` macro now implementing a type checker,
an interpreter (not implemented as a macro) for the meta-language
covering all features is added.

`source` is changed such that it adheres to the new rules and passes
type checking, but existing problems where the semantics as implemented
don't match the intended one are not yet fixed.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>