add a command-line shell for `LangDef` queries (#121)
## Summary

Add a simple command-line shell to query `LangDef`s with and execute
their functions and relations. The source language is currently
hardcoded as the language to query.

## Details

The shell is intended to aid with language development, by providing a
convenient way to test relations and functions with various input.

For ease of implementation, commands are provided as S-expressions.
Result values are also rendered as S-expressions. Constructors for
built-in meta-language terms (e.g., `true`, `false`, tuples, etc.) are
prefixed with a colon to distinguish them from user-defined type
constructors.

In addition, to showing the result of a function/relation application,
the shell also shows the interpreter-generated trace of which
relations.

### Additional Changes

* move the line-buffered stream and S-expression parsing implementations
  from the `repl` to the new `sexpstreams` module (`queryshell` also
  needs both)
* parse colon-prefixed symbols as S-expression symbols in the `parse`
  iterator, instead of reporting an error
* collect the used rules in traces generated by the interpreter