test the reference implementation (#120)
## Summary

Also run the `expr` tests using the source language reference
implementation `languages/source.nim`, to make sure that the
reference/specification implements/describes the desired semantics.

## Details

As a temporary solution, a standalone test orchestrator is used that
scans the `expr` directory for test files. For each test file, the
spec is extracted and (the relevant bits of it) parsed, with the test
content then parsed into a representation understood by the meta-
language interpreter.

The processing pipeline is: desugar -> type-check -> evaluate. Stages
are enabled/disabled based on the test specification. Failing tests
don't result in an orchestrator error, so that they can be fixed
incrementally; the list of known issues is hard-coded and needs to be
updated manually once more tests start to succeed.

There are many issues with the reference implementation as is, and
therefore not all tests can be run successfully at the moment.

A few small changes/additions are made to `source`, so that at least
some tests work:
* the `cstep` relation is added to `source`, which is the transitive
  closure of `step`. It's used for fully reducing (i.e., evaluating)
  an expression
* an infinite recursion with the `<:=` relation is fixed
* a catch-call branch is added to `desugar` so that the function
  doesn't fail for unhandled input

Finally, a dangling reference caused by a parameter aliasing violation
with the `mitems`'s iterator usage in `interpretRelation` is fixed.