queryshell: decouple from `languages/source.nim`
Summary
=======

Make the `queryshell` a standalone program that is not hardcoded to use
the language definition from `languages/source.nim`. The module
providing the definition must be imported at run-time, via the new
`import` command.

Details
=======

The `import` command takes a relative module path as input. It works by
compiling the provided module into a temporary dynamic library
exporting the `LangDef` instance (which at the moment must be named
`lang`) via a procedure, which is then loaded into the shell program.

This approach is chosen over the one detailed in the todo comment
because it doesn't require serialization and subsequent
deserialization, making it both easier to implement and more efficient.

The `LangDef` contains literal strings, which have to be copied to the
main program's heap memory prior to unloading the library
(`prepareMutation` takes care of this), otherwise the string's payload
pointers would start to dangle.

If no definition is loaded when a command requiring one is used, an
error is reported.