use arbitrary-precision integers in the meta language  (#130)
## Summary

* add a native `Bignum` implementation (`bignums.nim`)
* use `Bignum` for all arithmetic/comparison operations in the
  interpreter

## Details

A NimSkull-native `Bignum` implementation is used because:
* porting it to the source language in the future will be easier / can
  be automated
* the repository stays self-contained

Two explicit goals the implementation were that it's fully usable at
compile-time (in the VM) and that zero-initialized `Bignum`s are valid.
The implementation tries to strike a balance between complexity and
efficiency, usually preferring a simpler over a more efficient
implementation.

Using two's complement representation gets around having an extra sign
flag, which would increase the size of `Bignum`s by 8 byte (on a 64-bit
target).

Tests based on the `int128` tests are added for `Bignum`, with new
tests added for the various overflow/carry/extension scenarios.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>