lang: add explicit type widening (`As`) (#122)
## Summary

Add the `(As x typ)` expression to the source language, which is used
to explicitly widen the type of a value. `x` having a type that is
equal to `typ` is not an error.

## Details

`As` takes an expression and a type, fitting the former to the latter
where possible, reporting an error if not. At the moment, it's main
use is constructing values of `UnionTy` type, but in the future, it'll
also be useful for guiding type inference.

### Implementation

* lower `As` directly into a type conversion through `fitExpr` 
* fix error correction for fitting errors producing illformed `Expr`s,
  by keeping or discarding the trailing expression depending on the
  target type

Tests are added for `As`, and existing tests for union types now use
`As` instead of helper procedure for creating union values, greatly
simplifying the affected tests.