langdefs: create anonymous types for nested constructors
Nested constructors are now treated as proper constructors (with their
own symbol), by creating and assigning to them a dedicated anonymous
data type. This makes standalone usage of nested constructors possible,
which is something needed by (or at least useful for) the record
definition.

In addition, the `typeConstr` now aborts with a proper user-facing
error when there exists no constructor with the given name.