skully: fix codegen for setup of temporaries (#140)
## Summary

* fix code generator bug with skully, leading to crashes of the
  compiled code
* update to the `0.1.0-dev.21449` NimSkull compiler

## Details

### Skully

The `if` expressions emitted for `and` and `or` in the MIR currently
open no scope, meaning that the lifetime of temporaries created within
extend into the `if`'s scope.

Skully's code generator emits the location only for `def` statements,
and thus locations for which the `def` is in an unscoped `if` are in an
undefined state when the `if` is not entered, potentially leading to
memory corruption or access violations when using the location's
content (e.g., during destruction).

A MIR pass is added that moves moves the problematic `def`s (but not
the assignment) to right before the unscoped `if`, making all usages of
a location are dominated by the location's `def`.

### Misc

* bump the required NimSkull compiler version; the most recent release
  fixes a NimSkull bug that the above `skully` fix triggers
* bump the Ubuntu runner image to `22.04`; the recent NimSkull builds
  don't run on older images anymore (due to a too old glibc version)