lang: add dynamic tuple indexing (#135)
## Summary

Extend `At` to also allow tuple values as the first operand. If the
tuple is homogeneous, the expression has the type of the tuple element,
otherwise it's a union of all possible element types.

## Details

* add a typing rule for `At` with tuples
* add reduction rules for `At` and `With` covering tuples
* implement `At` support for tuple values in `source2il`
* fix a bug with `source2il.align` that resulted in the wrong value
  being returned (the mask was computed incorrectly)

`At` with a tuple operand never yields a mutable lvalue, even if the
input is one, as it'll make the future initial implementation of
pattern matching easier, but this decision is subject to change.