lang: add float multiplication and division (#141)
## Summary

Add the `*` and `/` operators for float values. Their operational
semantics are defined according to the IEEE 754 standard.

## Details

Both the definition in `source` and implementation in `source2il` are
straightforward. Tests are added to make sure multiplication and
division works, especially when involving non-finite values.

Integer and floating-point division are different enough to warrant
using separate identifiers (`/` vs. `div`). This also makes it possible
to add a `/` overload accepting integer values, in the future.

Division is defined in terms of a rational number division + to-float
conversion, as this is easier to understand than a division algorithm
using only integer numbers.

### Misc

Fix a small issue with the float addition overflow tests; they were
using constants that are outside the finite 64-bit floating point range,
meaning that the tests were effectively testing `inf` + finite. The
constants are changed to the largest possible finite values.