lang: add array type (#142)
## Summary

Add an array type to the source language, which represents a
homogeneous fixed-size vector type. Array values are constructed with
`ArrayCons`.

## Details

Arrays have significant overlap with homogeneous tuples, to the point
that they are essentially isomorphic (but they aren't considered as
such by the language, at the moment), most likely requiring a rework in
the near future.

For the moment, `ArrayTy` provides a more convenient way (when compared
to `TupleTy`) to express homogeneous vectors with many elements.
Compared to `TupleCons`, `ArrayCons` requires the elements' types to be
homogeneous, or that they at least have a common ancestor type.