cps: use tail calls for all continuations
Since tail calls require there to be no trailing implicit cleanup code,
the code in `interpreter` needs some adjustments and restructuring to
allow all continuation to be tail called.

This means using `sink` whenever sensible and manually destroying
locals/parameters. While quite noisy, it has the benefit of reducing
lifetimes overall, and thus also peak memory usage.

Finally, the inner procedures in `interpreter` are all changed to
`.tailcall`, so that the tail-call chain is not broken.