langdefs: fix a few type system issues
* fix the placeholder being typed as `all` instead of `void`
* fix branch pattern checking requring the matched expression having
  a type that's a subtype of or equal to the type of the pattern
  (instead of the other way around)
* fix the same issue for `where` predicates

`where` is effectively a run-time type check -- statically requiring
the left- and right-hand side to have matching types doesn't make much
sense and prevents some valid uses cases (which are now present in
`source`, e.g., in the `E-letrec` rule).

Making sure that there is at least overlap between the two sets would
be a nice-to-have, but the type-checker's pattern matching is not
developed enough to support this, at the moment.