langdefs: replace `tkData` with `tkSum`
A user-defined data type is now a sum type, with every constructor
pattern being its own type. This is much more accurate, and it also
allows for a slightly simpler implementation.

To implement the type matching in the interpreter, the type matching is
moved into its own procedure, so that it can self-recurse without
having to construct a temporary pattern node.