improve the meta-language's type system (#151)
## Summary

* assign a dedicated type to constructor patterns in a `typ` declaration
* distinguish between `A` and `A()` (value vs. nullary construction)
* slightly clean up type checking and speed it up

## Details

The type checker and type system are changed as follows:
* patterns representing types are now treated as such, using the
  `tkPat` type kind
* distinguish between `A` and `A()`. Both are now disjoint terms
* the `tkData` type kind is removed. `typ` and `subtype` declaration
  now create (potentially recursive) sums of pattern types

The changes allow for more narrow typing of value constructions.
Previously, they had to be typed as the *data type* matching the
construction, which is unnecessarily broad and necessitated `or` types
in order to be able to resolve ambiguity (i.e., when a construction
is covered by more than one data type).

As a side-effect of the more narrow typing, the type checker has to run
pattern matching less often, which significantly speeds up type
checking.

What's still missing is type relationship computation for pattern types.

Finally:
* the construction/value translation logic in `spectest` now that `A`
  and `A()` are different
* `queryshell.getDefault` is updated to handle the new `tkPat` type kind
* the pattern matching logic for the `interpreter` is updated to handle
  the new `tkPat` type kind