interpreter: use tail-calls for list matching (#152)
## Summary

Make the routines involved in list pattern matching `.tailcall`
routines. Usage of complex list patterns can no longer cause a stack
overflow (by itself).

## Details

So that tail-calling is possible, all list matching routines now fully
use continuation-passing style, instead of a mixture between it and
direct style.

As a consequence, repetition matching becomes more complex, as it can
no longer inspect the return value of the "rest". Instead, it needs
access to both the "rest" *and* the return continuation of the "rest",
so that it can logically intercept the return.

The way the implementation now works means that a repetition pattern
nested in a group no longer works, but this was never explicitly
supported anyway and it's also disallowed by the meta-language's
semantic checker.