phy: instrument pass execution
Summary
=======

Add the `--measure` switch to `phy`, which allows dumping internal
timings gathered for various steps. This is intended as a development
and optimization aid.

Details
=======

* add the `tracing` module, which implements a simple time-stamped-
  event log collection facility (`EventLog`)
* use `EventLog` to collect the start and end times for each internal
  pass, as well as `source2il` and VM code generation
* the timings are gathered from the log, rendered as a CSV list, and
  dumped to the standard output

To continue supporting compiling `phy` with `skully`, instrumentation is
disabled (at compile time), as the `times` and `monotimes` standard
library modules cannot be imported when using `--os:any` (which `skully`
always does).