ci: read NimSkull version from file (#163)
## Summary

The target NimSkull version is now supplied via a file, so that it's
easier to spot and there's only a single place it needs to be updated.

## Details

So that the version only needs to be read once, a new CI job
(`context`) is added, which provides the required NimSkull version as
an output to other jobs.