koch: replace `single` and `all` with `build` (#169)
## Summary

Replace the `single` and `all` commands with the more intuitive `build`
command. `build all` builds all programs (including `skully`), whereas
`build all-ws` builds all programs *except* `skully`.

## Details

The build logic in `skully` is restructured to be more generic. To make
lookup easier, the list of programs is now defined using a table, not
a sequence, with the program name used as the key.

For performance reasons, CI continues to build and test `skully`
separately from the other programs.