langs: add an IL with pointer types
It extends the `L0` language and is the basis for the `L1` language.

The `L3` is modified such that the new `Offset` is removed again. While
there's nothing preventing `Offset` from working in the `L3` and above,
it's better if higher-level ILs don't have access to pointer arithmetic
for now.