source2il: fix crash with copying records
Summary
=======

Fix the compiler crashing when copying a record that has a seq value as
a member.

Details
=======

The implementation for non-trivially-copyable records was missing,
causing the `unreachable` to trigger.

In order to prevent similar bugs going forward, the case statement for
producing the copy implementation is made exhaustive, using the new
`AllTypes` set. If a new aggregate type is added in the future, a
compile-time error will be reported if the new type is not handled.