source2il: fix crash with copying records (#176)
## Summary

Fix the compiler crashing for code where a record having a seq member
is copied.

## Details

The implementation for non-trivially-copyable records was missing,
causing the `unreachable` to trigger.

In order to prevent similar bugs going forward, the case statement for
producing the copy implementation is made exhaustive, using the new
`AllTypes` set. If a new aggregate type is added in the future, a
compile-time error will be reported if the new type is not handled.