source2il: fix crashes with erroneous code (#177)
## Summary

Fix the compiler crashing for some instances of semantically
invalid code.

## Details

Error types reached into `typeToString`, triggering the `unreachable`
assertion and thus crashing the compiler.

* change `typeToString` to render errors as `?`. There are valid cases
  for an error type being rendered, such as when it's part of a
  composite type
* improve error recovery for tuple analysis. An element being an error
  no longer short-circuits analysis by returning an error
* use a single procedure for handling type unification (`unify`). This
  fixes unnecessary follow-up errors being reported for an `If` where
  unification fails due to one type being an error
* don't report unnecessary follow-up errors in `fitExprStrict`

Tests are added to make sure code that previously crashed the compiler
no longer does. The change to `typeToString` itself is enough to fix the
crash -- the other changes are to not report useless diagnostics.